requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.8.2  # https://github.com/evansd/whitenoise
redis==5.2.1  # https://github.com/redis/redis-py
hiredis==3.1.0  # https://github.com/redis/hiredis-py
chardet
pandas
django-jazzmin==3.0.1
django-widget-tweaks#==1.4.11  # https://github.com/jazzband/django-widget-tweaks

# Django
# ------------------------------------------------------------------------------
django==4.2.21  # pyup: < 5.1  # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.3.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.3  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
djangorestframework==3.14.*
